relax requirements
This package is far too aggressive in its listed requirements: Given
that other packages depend on it, it should simply list what it needs,
and let _them_ handle any further breakage, as required.

As an example, `oio_rest` currently requires an `itsdangerous` that's
newer than the 1.0 this package explicitly says isn't good enough.